ganti publisher jadi owner

+++ src/Applications/use_case/CommentUseCase.js
@@ -16,21 +16,21 @@ class CommentUseCase {
 
   async deleteComment (useCasePayload) {
     this._verifyPayload(useCasePayload)
-    const { commentId, threadId, publisher } = useCasePayload
+    const { commentId, threadId, owner } = useCasePayload
     await this._threadRepository.checkAvailabilityThread(threadId)
     await this._commentRepository.checkAvailabilityComment(commentId)
-    await this._commentRepository.verifyCommentpublisher(commentId, publisher)
+    await this._commentRepository.verifyCommentowner(commentId, owner)
     await this._commentRepository.deleteComment(commentId)
   }
 
   _verifyPayload (payload) {
-    const { commentId, threadId, publisher } = payload
+    const { commentId, threadId, owner } = payload
 
-    if (!commentId || !threadId || !publisher) {
+    if (!commentId || !threadId || !owner) {
       throw new Error('DELETE_COMMENT_USE_CASE.NOT_CONTAIN_NEEDED_PROPERTY')
     }
 
-    if (typeof commentId !== 'string' || typeof threadId !== 'string' || typeof publisher !== 'string') {
+    if (typeof commentId !== 'string' || typeof threadId !== 'string' || typeof owner !== 'string') {
       throw new Error('DELETE_COMMENT_USE_CASE.NOT_MEET_DATA_TYPE_SPECIFICATION')
     }
   }
